services/incident-service/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
httpx==0.28.1
redis==5.2.1
pydantic-settings==2.13.1
python-jose[cryptography]==3.3.0
